+++ src/command.ts
@@ -1,7 +1,12 @@
 import { GetHelp } from "./errors.ts";
 import { getFlagsetHelp, getFlagsetParser } from "./flagset.ts";
-import type { Command, CommandFn, Flagset } from "./types.ts";
-import type { Writer } from "@std/io";
+import type {
+  Command,
+  CommandFn,
+  Flagset,
+  StringOutput,
+  Writer,
+} from "./types.ts";
 
 export function command<VV>(
   opts: {
@@ -20,20 +25,26 @@ export function command<VV>(
   };
 }
 
+export function makeStringOutput(output: Writer): StringOutput {
+  const encoder = new TextEncoder();
+  return (msg: string) => output.write(encoder.encode(msg));
+}
+
 export async function runCommand<VV>(
+  // TODO: add name here or in command object, which may become names path in multicommand
   cmd: Command<VV>,
   args: string[],
   output: Writer,
 ): Promise<number> {
-  const encoder = new TextEncoder();
-  const write = async (msg: string) => await output.write(encoder.encode(msg));
+  const write = makeStringOutput(output);
 
   try {
     const params = cmd.parse(args);
     const result = await cmd.execute(params, write);
     return result;
   } catch (err) {
-    await write(GetHelp(err));
+    await write(cmd.help());
+    await write("\n" + GetHelp(err));
     return 999;
   }
 }